requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
pydantic==2.5.0
pytz==2023.3
openpyxl==3.1.5
reportlab==4.0.9
